requirements.txt
asgiref==3.4.1
Django==3.2.8
django-filter==21.1
django-restful-admin==1.1.3
djangorestframework==3.13.1
djangorestframework-simplejwt==5.0.0
drf-base64==2.0
gunicorn==20.1.0
Pillow==9.0.0
PyJWT==2.3.0
pytz==2021.3
sqlparse==0.4.2
